Move command-line flags to a new file flags.go
